requirements.txt
openai
flask
flask-sqlalchemy
python-dotenv
werkzeug
flask-login
psycopg2-binary
jsonify
supabase
langchain
langchain-community
langchain-openai
langgraph
lark
gutenbergpy